new components, change in update
added the camera, transfrom, and modelcomponent classes. modelcomponent still needs to be filled in transformcomponent still needs 1 function filled in. removed the pass of deltaTime in the update/fixedupdate i will give the gamecontext a time value later that holds the deltaT and fixedT. made a tiny change to my if check in the inputmanager update function.